+++ giottotime/feature_extraction/tests/test_standard.py
