+++ giottotime/feature_extraction/trend.py
